cleanup folders. insert 4.
