+++ challenge_4/UnitTest4.cs
@@ -1,23 +1,37 @@
+using System;
 using System.Threading;
 using System.Threading.Tasks;
 using Xunit;
+using Xunit.Abstractions;
 
 namespace cancellation_practice.challenge_4
 {
     public class UnitTest4
     {
-        /// <summary>
-        /// In this test, we have a cancellationtoken we can not cancel
-        /// WHAT TO DO? 
-        /// The idea here is to show some ideas that might make it a bit more clear
-        /// How the polly cancellation tokens are working
-        /// </summary>
+        private readonly ITestOutputHelper _testOutputHelper;
+
+        public UnitTest4(ITestOutputHelper testOutputHelper)
+        {
+            _testOutputHelper = testOutputHelper;
+        }
+
         [Fact]
         public async Task Test4()
         {
+            var timer = Task.Delay(1000);
+            var cancellationTokenSource = new CancellationTokenSource();
+            var cancellationToken = cancellationTokenSource.Token; 
+            cancellationTokenSource.Cancel(); 
+
             var candyFetcher = new CandyFetcher();
-            var candy = candyFetcher.FetchCandy(CancellationToken.None);
-            await Assert.ThrowsAsync<TaskCanceledException>(async () => await candy);
+            var candy = candyFetcher.FetchCandy(cancellationToken);
+            
+            await Assert.ThrowsAsync<OperationCanceledException>(async () =>
+            {
+                var result = await candy;
+                _testOutputHelper.WriteLine("You got candy:" + result.Amount);
+            });
+            Assert.False(timer.IsCompleted);
         }
     }
 }
